Add files via upload
First upload of basic proof of concept for an Arduino Alarm System. It consists of an Arduino Uno clone, an input button to simulate arming and disarming, a PIR sensor, and a LED to simulate a siren output. So far, the system will detect an arming and only react to inputs when armed triggering the "siren". The "siren" will only trigger once during an arming, and is reset when  disarmed and armed again.